// src/lib/server/types/db/runs/dfa/dfaSolo.ts
export interface DfaSoloRunDbModel {
	RunID: string;
	PlayerID: string;
	RunCharacterName: string;
	Patch: string;
	Rank: string;
	Time: string;
	Buff: string;
	MainClass: string;
	SubClass: string;
	WeaponInfo1: string;
	WeaponInfo2: string;
	WeaponInfo3: string;
	WeaponInfo4: string;
	WeaponInfo5: string;
	WeaponInfo6: string;
	Link: string;
	Notes: string;
	ModNotes: string;
	SubmitterID: string;
	SubmissionTime: string;
	VideoTag: string;
	Drill: string;
	PlayerName: string;
	CharacterName: string;
	Description: string;
	Youtube: string;
	Twitch: string;
	Twitter: string;
	Discord: string;
	PreferredName: string;
	Server: string;
	Ship: string;
	Flag: string;
	BackgroundType: string;
	BackgroundImage: string;
	BackgroundColor: string;
	NameType: string;
	NameColor1: string;
	NameColor2: string;
}